Completed learning AI

+++ pkg/db/sql/training/create_table_results.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS results
 (
     id INT NOT NULL AUTO_INCREMENT,
-    val DOUBLE,
+    val INT,
     training_id INT,
     PRIMARY KEY(id)
 );
